[TM-1959] feat: refactor Disturbance processor and DTOs for improved entity handling and add Invasive entity support

+++ libs/common/src/lib/events/entity-status-update.event-processor.ts
@@ -16,6 +16,7 @@ import { get, isEmpty, map } from "lodash";
 import { Op } from "sequelize";
 import { STATUS_DISPLAY_STRINGS } from "@terramatch-microservices/database/constants/status";
 import { Disturbance } from "@terramatch-microservices/database/entities/disturbance.entity";
+import { Invasive } from "@terramatch-microservices/database/entities/invasive.entity";
 
 export type StatusUpdateModel = LaravelModel & StatusModel & FeedbackModel;
 
@@ -56,6 +57,7 @@ export class EntityStatusUpdate extends EventProcessor {
     const entity = this.model as EntityModel;
 
     if (entity instanceof Disturbance) return;
+    if (entity instanceof Invasive) return;
 
     await Action.for(entity).destroy({ where: { type: "notification" } });
